requirements.txt
openai
streamlit
python-dotenv
#audio-recorder-streamlit
#streamlit_float
